+++ src/main/java/ctbrec/ui/JavaFxModel.java
@@ -1,6 +1,9 @@
 package ctbrec.ui;
 
+import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.ExecutionException;
 
 import ctbrec.Model;
 import javafx.beans.property.BooleanProperty;
@@ -16,7 +19,9 @@ public class JavaFxModel extends Model {
 
     public JavaFxModel(Model delegate) {
         this.delegate = delegate;
-        setOnline(delegate.isOnline());
+        try {
+            onlineProperty.set(Objects.equals("public", delegate.getOnlineState(true)));
+        } catch (IOException | ExecutionException e) {}
     }
 
     @Override
@@ -57,17 +62,6 @@ public class JavaFxModel extends Model {
     @Override
     public void setTags(List<String> tags) {
         delegate.setTags(tags);
-    }
-
-    @Override
-    public boolean isOnline() {
-        return delegate.isOnline();
-    }
-
-    @Override
-    public void setOnline(boolean online) {
-        delegate.setOnline(online);
-        this.onlineProperty.set(online);
     }
 
     @Override